Migrate toolbar to lvgl-module
