+++ Tactility/Source/Tactility.cpp
@@ -25,7 +25,7 @@
 
 #include <crypt/module.h>
 #include <lvgl/module.h>
-#include <Tactility/lvgl/Toolbar.h>
+#include <lvgl/widgets/toolbar.h>
 #include <tactility/concurrent/thread.h>
 #include <tactility/drivers/audio_stream.h>
 #include <tactility/drivers/display.h>
@@ -356,7 +356,8 @@ static void stopAppFromToolbar(lv_event_t*) {
 }
 
 static void onLvglStarted() {
-    lvgl::toolbar_configure({ .navActionCallback = stopAppFromToolbar });
+    ToolbarConfig toolbar_config = { .nav_action_callback = stopAppFromToolbar };
+    lvgl_toolbar_configure(&toolbar_config);
 
     addService(service::gui::manifest);
     addService(service::statusbar::manifest);
