+++ Tactility/Source/app/fileselection/View.cpp
@@ -7,7 +7,7 @@
 #include <Tactility/file/File.h>
 #include <Tactility/Platform.h>
 #include <Tactility/lvgl/LvglSync.h>
-#include <Tactility/lvgl/Toolbar.h>
+#include <lvgl/widgets/toolbar.h>
 #include <tactility/check.h>
 #include <tactility/log.h>
 
@@ -181,8 +181,8 @@ void View::init(lv_obj_t* parent, Mode mode) {
     lv_obj_set_flex_flow(parent, LV_FLEX_FLOW_COLUMN);
     lv_obj_set_style_pad_row(parent, 0, LV_STATE_DEFAULT);
 
-    auto* toolbar = lvgl::toolbar_create(parent, "Select File");
-    navigate_up_button = lvgl::toolbar_add_image_button_action(toolbar, LV_SYMBOL_UP, &onNavigateUpPressedCallback, this);
+    auto* toolbar = lvgl_toolbar_create(parent, "Select File");
+    navigate_up_button = lvgl_toolbar_add_image_button_action(toolbar, LV_SYMBOL_UP, &onNavigateUpPressedCallback, this);
 
     auto* wrapper = lv_obj_create(parent);
     lv_obj_set_width(wrapper, LV_PCT(100));
