+++ Tactility/Source/app/webserversettings/WebServerSettings.cpp
@@ -192,7 +192,7 @@ public:
         lv_obj_t* toolbar = lvgl::toolbar_create(parent, app);
 
         // Web Server Enable toggle
-        switchWebServerEnabled = lvgl::toolbar_add_switch_action(toolbar);
+        switchWebServerEnabled = lvgl_toolbar_add_switch_action(toolbar);
         if (wsSettings.webServerEnabled) {
             lv_obj_add_state(switchWebServerEnabled, LV_STATE_CHECKED);
         }
